Move benchmark into repo owned files
